Update money types to decimal
Fixing rounding and addition issues by updating all floats to decimals

+++ App/HardTotals.cs
@@ -9,6 +9,6 @@ namespace BT_COMMONS.App;
 
 public class HardTotals
 {
-    public Dictionary<TransactionTender, float>? Tender { get; set; }
-    public Dictionary<TransactionType, float>? Type { get; set; } 
+    public Dictionary<TransactionTender, decimal>? Tender { get; set; }
+    public Dictionary<TransactionType, decimal>? Type { get; set; } 
 }
